Extract git.GitEnv and github.GithubEnv into envs module
commit-id:5140ecd5
